+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ### Added
 
-- Cursor changes when over a button
-- It's possible to specify widths on the FormFieldLine
+- Cursor changes when over a button.
+- It's possible to specify widths on the FormFieldLine.
+- ButtonLine component supports `align` property.
+- Added DescriptiveHeader component.
 
 ### Fixed